Figured it out!
The trick was to add to the header search path:

    "$(DERIVED_SOURCES_DIR)/../../MixedLanguageTesting.build/DerivedSources"

That source directory gets populated when the main target is built.
It has the -Swift.h Swift->Obj-C bridging header we're trying to import.

Two ugly things:

- Though Xcode doesn't show it in the expansion preview, this winds up
  dumped under an Intermediate-noindex folder, which is a sibling of
  the parent of $(BUILD_DIR). So building down from the top,
  rather than backing up out of our DERIVED_SOURCES_DIR, doesn't work so
  well.
- I'd like to see if there's a way to get at the main app target's
  PRODUCT_NAME directly, so that I don't have to hardcode it in there.

Those are cosmetic issues, though. The former entirely, the latter as
long as you're not renaming your app target all the time.